pyproject.toml
[tool.poetry]
name = "vebeg_scraper"
version = "0.1.0"
description = ""
authors = ["David <david.scheiklies@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.1"
pyyaml = "^5.3.1"
prometheus_client = "^0.8.0"
schedule = "^0.6.0"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.5.3"
mypy = "^0.782"
black = "^19.10b0"
pre-commit = "^2.6.0"
flake8 = "^3.8.3"
rope = "^0.17.0"
debugpy = "^1.0.0-rc.2"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[tool.isort]
profile = "black"
src_paths = [ "test", "{{cookiecutter.project_slug}}"]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests"
]
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


